Move main file to app/ subdir.
This works around one of the ways in which the App Engine SDK is
brain-dead, namely that it searches _recursively_ for Go files to
compile into the main program. This clashes, of course, with having
sensible subpackage splits. The workaround was to add all the package
subdirs into a `nobuild_packages` regexp. This nonsense can now go.