feat: show failed dependencies in package results.
This table existed in the App Engine version and was temporarily removed,
now it's back! Unlike the "packages broken by this one", the table is
not AJAX, for various reasons. The main reason is that we already have
the list of failed dependencies, so no need for another round trip.