Add all non-OK dependencies to failedDeps.
FailedDeps used to be only failed, but there can be indirectly failed
dependencies as well. This is a bug that has been in the app for years!

I hope that this addressed the discrepancies in terms of the number of failed
dependencies in general.

There is still a bug around the "Breaks" counter. Left a TODO for now.

This will eventually be obsoleted by the fix for #60.