Switch from Google Chart API to bundled charts.js
This is for the donut chart in the bulk build overview page that
gives the number of OK and failed packages.

chart.js uses canvas instead of svg, which is faster in theory. It is
also not deprecated, contrary to the Google Chart API. And the app no
longer needs the Google JSAPI at all!

Fixes issue #46