bulk: change algorithm for failed dependencies
A package is now only considered to "break" another package if it is
Failed or Prefailed, not IndirectFailed or IndirectPrefailed.

So if a depends on b depends on c, and c failed to build, a's failed
dependency is c, not b. (There is a test for exactly this!)

This matches how pbulk does it and is more correct and useful.